Fix connect tab removing filters on refresh
- Connect tab has been refactored into a separate file as the logic had
  nothing in common with the receiver show view.
- To prevent unnecessary recreations of the connect tab, the component
  is stored as state and only recreated when the receiver data is
  changed.
- The render prop has been removed from the tab Route components.
