+++ Development/src/pages/receivers/ConnectionManagementTab.js
@@ -1,0 +1,199 @@
+import React, { Fragment, useEffect, useState } from 'react';
+import Link from 'react-router-dom/Link';
+import {
+    Card,
+    CardContent,
+    Table,
+    TableBody,
+    TableCell,
+    TableHead,
+    TableRow,
+} from '@material-ui/core';
+import {
+    Loading,
+    ReferenceField,
+    TextField,
+    TitleForRecord,
+    linkToRecord,
+} from 'react-admin';
+import get from 'lodash/get';
+import useGetList from '../../components/useGetList';
+import FilterPanel, {
+    BooleanFilter,
+    StringFilter,
+} from '../../components/FilterPanel';
+import QueryVersion from '../../components/QueryVersion';
+import ChipConditionalLabel from '../../components/ChipConditionalLabel';
+import ActiveField from '../../components/ActiveField';
+import ConnectButtons from './ConnectButtons';
+import PaginationButtons from '../../components/PaginationButtons';
+import { ReceiversTitle } from './ReceiversShow';
+
+const ConnectionManagementTab = ({ receiverData, basePath }) => {
+    const [filter, setFilter] = useState({
+        transport: get(receiverData, 'transport'),
+    });
+    const [paginationURL, setPaginationURL] = useState(null);
+    const { data, loaded, pagination } = useGetList({
+        basePath,
+        filter,
+        paginationURL,
+        resource: 'senders',
+    });
+
+    // receiverData initialises undefined, update when no longer null
+    useEffect(() => {
+        if (receiverData !== null)
+            setFilter({ transport: get(receiverData, 'transport') });
+    }, [receiverData]);
+
+    if (!loaded) return <Loading />;
+
+    const nextPage = label => {
+        setPaginationURL(pagination[label]);
+    };
+
+    return (
+        <Fragment>
+            <TitleForRecord
+                record={receiverData}
+                title={<ReceiversTitle record={receiverData} />}
+            />
+            <Card>
+                <CardContent>
+                    <FilterPanel
+                        data={data}
+                        filter={filter}
+                        setFilter={setFilter}
+                    >
+                        <StringFilter source="label" label="Sender Label" />
+                        <StringFilter
+                            source="description"
+                            label="Sender Description"
+                        />
+                        {QueryVersion() >= 'v1.2' && (
+                            <BooleanFilter
+                                source="subscription.active"
+                                label="Sender Active"
+                            />
+                        )}
+                        <StringFilter source="id" label="Sender ID" />
+                    </FilterPanel>
+                    <Table>
+                        <TableHead>
+                            <TableRow>
+                                <TableCell
+                                    style={{
+                                        paddingLeft: '32px',
+                                    }}
+                                >
+                                    Sender
+                                </TableCell>
+                                {QueryVersion() >= 'v1.2' && (
+                                    <TableCell>Active</TableCell>
+                                )}
+                                <TableCell>Flow</TableCell>
+                                <TableCell>Format</TableCell>
+                                {QueryVersion() >= 'v1.1' && (
+                                    <TableCell>Media Type</TableCell>
+                                )}
+                                <TableCell>Connect</TableCell>
+                            </TableRow>
+                        </TableHead>
+                        <TableBody>
+                            {data.map(item => (
+                                <TableRow
+                                    key={item.id}
+                                    selected={
+                                        get(receiverData, 'id') === item.id
+                                    }
+                                >
+                                    <TableCell component="th" scope="row">
+                                        {
+                                            // Using linkToRecord as ReferenceField will
+                                            // make a new unnecessary network request
+                                        }
+                                        <Link
+                                            to={`${linkToRecord(
+                                                '/senders',
+                                                item.id
+                                            )}/show`}
+                                            style={{
+                                                textDecoration: 'none',
+                                            }}
+                                        >
+                                            <ChipConditionalLabel
+                                                record={item}
+                                                source="label"
+                                                label="ra.action.show"
+                                            />
+                                        </Link>
+                                    </TableCell>
+                                    {QueryVersion() >= 'v1.2' && (
+                                        <TableCell>
+                                            <ActiveField
+                                                record={item}
+                                                resource="senders"
+                                            />
+                                        </TableCell>
+                                    )}
+                                    <TableCell>
+                                        <ReferenceField
+                                            record={item}
+                                            basePath="/flows"
+                                            label="Flow"
+                                            source="flow_id"
+                                            reference="flows"
+                                            link="show"
+                                        >
+                                            <ChipConditionalLabel source="label" />
+                                        </ReferenceField>
+                                    </TableCell>
+                                    <TableCell>
+                                        <ReferenceField
+                                            record={item}
+                                            basePath="/flows"
+                                            label="Flow"
+                                            source="flow_id"
+                                            reference="flows"
+                                            link={false}
+                                        >
+                                            <TextField source="format" />
+                                        </ReferenceField>
+                                    </TableCell>
+                                    {QueryVersion() >= 'v1.1' && (
+                                        <TableCell>
+                                            <ReferenceField
+                                                record={item}
+                                                basePath="/flows"
+                                                label="Flow"
+                                                source="flow_id"
+                                                reference="flows"
+                                                link={false}
+                                            >
+                                                <TextField source="media_type" />
+                                            </ReferenceField>
+                                        </TableCell>
+                                    )}
+                                    <TableCell>
+                                        <ConnectButtons
+                                            senderData={item}
+                                            receiverData={receiverData}
+                                        />
+                                    </TableCell>
+                                </TableRow>
+                            ))}
+                        </TableBody>
+                    </Table>
+                    <br />
+                    <PaginationButtons
+                        disabled={!pagination}
+                        nextPage={nextPage}
+                    />
+                </CardContent>
+            </Card>
+        </Fragment>
+    );
+};
+
+export default ConnectionManagementTab;
